pyproject.toml
[project]
name = "loom-core"
version = "1.0.0"
description = "Loom Core daemon — personal knowledge fabric (FastAPI + SQLite + Obsidian)"
requires-python = ">=3.13,<3.14"
authors = [{ name = "Phani" }]
readme = "README.md"

dependencies = [
    "fastapi>=0.115,<0.116",
    "uvicorn[standard]>=0.32,<0.33",
    "pydantic>=2.9,<3.0",
    "sqlalchemy[asyncio]>=2.0,<2.1",
    "aiosqlite>=0.20,<0.21",
    "alembic>=1.13,<1.14",
    "python-ulid>=2.7,<3.0",
    "python-frontmatter>=1.1,<1.2",
    "mistune>=3.0,<4.0",
    "jinja2>=3.1,<4.0",
    "httpx>=0.27,<0.28",
    "anthropic>=0.97,<0.98",
    "sentence-transformers>=3.3,<4.0",
    "rapidfuzz>=3.10,<4.0",
    "apscheduler>=3.10,<4.0",
    "structlog>=24.4,<25.0",
    "typer>=0.13,<0.14",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
    "pytest-asyncio>=0.24,<0.25",
    "ruff>=0.7,<0.8",
    "mypy>=1.13,<2.0",
    # loom-mcp is imported in integration tests to exercise the seam.
    # The path source is declared in [tool.uv.sources] below.
    "loom-mcp",
]

[project.scripts]
loom = "loom_core.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/loom_core"]

[tool.hatch.metadata]
allow-direct-references = true

# ---------------------------------------------------------------------------
# uv — local path overrides for workspace siblings
# ---------------------------------------------------------------------------
[tool.uv.sources]
loom-mcp = { path = "../loom-mcp", editable = true }

# ---------------------------------------------------------------------------
# Ruff — linter + formatter
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py313"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "RUF", # ruff-specific
]
ignore = [
    "E501",  # line length handled by formatter
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["B011"]  # asserts in tests

# ---------------------------------------------------------------------------
# Mypy — strict type checking
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["alembic.*", "apscheduler.*", "frontmatter", "mistune.*", "sentence_transformers.*"]
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra --strict-markers"
markers = [
    "integration: marks integration tests (requires temp filesystem)",
    "external: marks tests that hit real external services (default skipped)",
    "visibility: marks visibility-filter regression tests (#079)",
    "operations_log: marks operations-log replay tests (#089)",
]
